support `mode` on physical machine chaos (#2866)
* fix: support mode on physical machine chaos


* chore: fix handleSwitchEnv type


* chore: update
